Shift 'AccordionContainer.store' to 'AccordionStore' and rename 'AccordionContainer' to 'AccordionContext'
