+++ src/AccordionContext/AccordionContext.spec.tsx
@@ -1,11 +1,7 @@
 import { mount } from 'enzyme';
 import * as React from 'react';
-import {
-    AccordionContainer,
-    Consumer,
-    Item,
-    Provider,
-} from './AccordionContainer';
+import { Item } from '../AccordionStore/AccordionStore';
+import { AccordionContainer, Consumer, Provider } from './AccordionContext';
 
 const DEFAULT_ITEM: Item = {
     uuid: 'foo',
