+++ src/AccordionItem/AccordionItem.tsx
@@ -1,9 +1,7 @@
 import { default as classnames } from 'classnames';
 import * as React from 'react';
-import {
-    AccordionContainer,
-    Item,
-} from '../AccordionContainer/AccordionContainer';
+import { AccordionContainer } from '../AccordionContext/AccordionContext';
+import { Item } from '../AccordionStore/AccordionStore';
 import { UUID } from '../ItemContainer/ItemContainer';
 
 type AccordionItemProps = React.HTMLAttributes<HTMLDivElement> & {
