+++ src/AccordionStore/AccordionStore.ts
@@ -8,7 +8,7 @@ export type Item = {
     expanded: boolean;
 };
 
-export default class AccordionContainerStore {
+export default class AccordionStore {
     public readonly items: Item[];
     public readonly allowMultipleExpanded: boolean;
     public readonly allowZeroExpanded: boolean;
@@ -27,7 +27,7 @@ export default class AccordionContainerStore {
         this.allowZeroExpanded = allowZeroExpanded;
     }
 
-    public readonly addItem = (newItem: Item): AccordionContainerStore => {
+    public readonly addItem = (newItem: Item): AccordionStore => {
         if (this.items.some((item: Item) => item.uuid === newItem.uuid)) {
             // tslint:disable-next-line:no-console
             console.error(
@@ -54,7 +54,7 @@ export default class AccordionContainerStore {
         });
     };
 
-    public readonly removeItem = (key: UUID): AccordionContainerStore => {
+    public readonly removeItem = (key: UUID): AccordionStore => {
         return this.augment({
             items: this.items.filter((item: Item) => item.uuid !== key),
         });
@@ -63,7 +63,7 @@ export default class AccordionContainerStore {
     public readonly setExpanded = (
         key: UUID,
         expanded: boolean,
-    ): AccordionContainerStore => {
+    ): AccordionStore => {
         if (this.isItemDisabled(key)) {
             return this;
         }
@@ -113,8 +113,8 @@ export default class AccordionContainerStore {
         items?: Item[];
         allowMultipleExpanded?: boolean;
         allowZeroExpanded?: boolean;
-    }): AccordionContainerStore => {
-        return new AccordionContainerStore({
+    }): AccordionStore => {
+        return new AccordionStore({
             items: this.items,
             allowMultipleExpanded: this.allowMultipleExpanded,
             allowZeroExpanded: this.allowZeroExpanded,
